Refactor authentication setup by moving authOptions to a separate module and updating import paths
