+++ src/app/lib/auth.ts
@@ -1,0 +1,56 @@
+import * as mongoose from 'mongoose';
+import bcrypt from 'bcrypt';
+import { NextAuthOptions } from 'next-auth';
+import CredentialsProvider from 'next-auth/providers/credentials';
+import GoogleProvider from 'next-auth/providers/google';
+import { MongoDBAdapter } from '@auth/mongodb-adapter';
+import clientPromise from '@/app/lib/mongoClient';
+import { User } from '@/models/User';
+
+export const authOptions: NextAuthOptions = {
+  secret: process.env.NEXTAUTH_SECRET,
+  adapter: MongoDBAdapter(clientPromise, {
+    databaseName: process.env.MONGODB_DB,
+  }),
+  session: {
+    strategy: 'jwt',
+    maxAge: 30 * 24 * 60 * 60,
+    updateAge: 24 * 60 * 60,
+  },
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID!,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+    }),
+    CredentialsProvider({
+      id: 'credentials',
+      name: 'Credentials',
+      credentials: {
+        email: { label: 'Email', type: 'email' },
+        password: { label: 'Password', type: 'password' },
+      },
+      async authorize(credentials, req) {
+        const email = credentials?.email;
+        const password = credentials?.password;
+
+        if (!email || !password) return null;
+
+        if (!process.env.MONGODB_URI || !process.env.MONGODB_DB) {
+          throw new Error(
+            'Missing env variables: "MONGODB_URI" Or "MONGODB_DB"',
+          );
+        }
+        mongoose.connect(process.env.MONGODB_URI, {
+          dbName: process.env.MONGODB_DB,
+        });
+        const user = await User.findOne({ email });
+        const passwordOk = user && bcrypt.compareSync(password, user.password);
+        if (passwordOk) {
+          return { id: user.id, email: user.email };
+        }
+
+        return null;
+      },
+    }),
+  ],
+};
