+++ src/app/lib/userInfos.ts
@@ -2,7 +2,7 @@ import * as mongoose from 'mongoose';
 import { User } from '@/models/User';
 import { getServerSession } from 'next-auth';
 import { UserInfo } from '@/models/UserInfo';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/app/lib/auth';
 
 
 export async function isAdmin() {
